backend/pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = ""
authors = ["Vud <vud150294@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
pre-commit = "^4.0.1"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
commitizen = "^3.29.1"
sqlalchemy = "^2.0.36"
pydantic = {extras = ["email"], version = "^2.9.2"}
bcrypt = "^4.2.0"
jwt = "^1.3.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.32.0"
alembic = "^1.13.3"
python-multipart = "^0.0.12"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
